refactor: Make `Distribution` a struct
Previously, `Distribution` was an enum with variants representing the
different archive formats. However, the inspector url field was
duplicated for each format. This change removes the duplication in the
fields.

Also adds the `Scan` trait, which the archive types now implement.

However, this change also introduces a minor performance hit, due to the
dynamic dispatch of the trait object, but this should be a worthwhile
tradeoff.