Test that ScanResult serializes into Mainframe response model
Test that the ScanResult enum properly serializes into a format that
Mainframe expects. Previously, `Result<T, E>` would serialize into
`{"Ok": ...}` if the Result was Ok, or if it was Err, then it would
serialize into `{"Err": ...}` which made Mainframe throw 422
Unprocessable Entity, since it was not expecting responses to be in that
format.