track authentication via cookie store
Instead of tracking the access token in a struct ourselves, we perform
an initial authentication to the root URL with the Cloudflare service
token ID and secret. The cloudflare proxy attaches a CF_Authorization
token in the cookies, which is then tracked by reqwest's built in cookie
store. Therefore, we don't need to accept an access_token argument for
every method, as it's implicitly included in the http_client argument.